fix(scraper): unifica detecção de bloqueio e protege a página de produto
A detecção de bloqueio vivia duplicada com markers divergentes em
http_client.is_blocked (listagem) e update_service.is_block_page (página de
produto). Além disso, a página de produto fazia session.get cru, sem rate
limiter nem backoff — a camada mais exposta a ban era a menos protegida.

- Novo módulo scrapper_mlb/block_detection.py concentra os markers; ambas as
  camadas passam a delegar a is_blocked(html).
- update_service reaproveita o rate_limiter e o backoff globais do
  http_client: espera antes do GET, backoff.error() em erro/bloqueio e
  backoff.success() em resposta válida; o retry agora usa backoff.wait()
  em vez de sleep fixo.

Refs #52
