+++ scrapper_mlb/product_page/service/update_service.py
@@ -1,28 +1,15 @@
 import os
-import time
 import requests
 from bs4 import BeautifulSoup
+from scrapper_mlb.block_detection import is_blocked
 from scrapper_mlb.config import HEADERS
+from scrapper_mlb.http_client import backoff, rate_limiter
 from scrapper_mlb.product_page.update_builder import build_product_page
 
 DEBUG_DIR = "debug_html"
 
 session = requests.Session()
 session.headers.update(HEADERS)
-
-
-def is_block_page(soup: BeautifulSoup) -> bool:
-    # Detecta bloqueio REAL
-    if soup.select_one(".account-verification-header"):
-        return True
-
-    if soup.select_one(".new-user-button"):
-        return True
-
-    if "Olá! Para continuar, acesse" in soup.get_text():
-        return True
-
-    return False
 
 
 def clean_url(url: str) -> str:
@@ -34,20 +21,24 @@ def _collect_once(url: str, produto_id: int):
     try:
         url = clean_url(url)
 
+        # 🔥 Mesmo rate limiter/backoff da listagem (politeness global)
+        rate_limiter.wait()
         response = session.get(url, timeout=15, allow_redirects=True)
 
         if response.status_code != 200:
             print(f"⚠️ Status inválido: {response.status_code}")
+            backoff.error()
+            return None
+
+        # 🔥 Detecção de bloqueio unificada (block_detection)
+        if is_blocked(response.text):
+            print(f"🚨 BLOQUEIO detectado para ID={produto_id}")
+            backoff.error()
             return None
 
         soup = BeautifulSoup(response.text, "html.parser")
-
-        # 🔥 Detecta bloqueio real
-        if is_block_page(soup):
-            print(f"🚨 BLOQUEIO detectado para ID={produto_id}")
-            return None
-
         page_data = build_product_page(soup)
+        backoff.success()
 
         # 🔥 Debug se não tiver preço
         if page_data.preco is None:
@@ -65,6 +56,7 @@ def _collect_once(url: str, produto_id: int):
 
     except requests.RequestException as e:
         print(f"⚠️ Erro de requisição: {e}")
+        backoff.error()
         return None
 
 
@@ -79,7 +71,9 @@ def collect_product_by_url(url: str, produto_id: int, retries: int = 1):
 
         if attempt < retries:
             print(f"🔁 Retry {attempt + 1}/{retries} para ID={produto_id}")
-            time.sleep(5)
+            # Espera com backoff exponencial (inflado pelo backoff.error()
+            # disparado no _collect_once em caso de erro/bloqueio).
+            backoff.wait()
 
     print(f"❌ Falha definitiva para ID={produto_id}")
     return None
